requirements.txt
pydantic
uvicorn
fastapi
python-dotenv
jose
pydantic-settings
bcrypt
passlib
python-jose
sqlalchemy
alembic
asyncpg
python-multipart
psycopg2
asyncpg
libpq-dev